modified:   backend/api/src/routes/me/submissions.rs 	modified:   backend/api/src/routes/modules/assignments/mark_allocator/put.rs 	modified:   backend/api/src/routes/modules/assignments/statistics/get.rs 	modified:   backend/api/tests/helpers/config_assert.rs 	modified:   backend/api/tests/routes/me/grades_tests.rs 	modified:   backend/api/tests/routes/me/submissions_tests.rs 	modified:   backend/api/tests/routes/modules/assignments/mark_allocator/get_test.rs 	modified:   backend/api/tests/routes/modules/assignments/mark_allocator/post_test.rs 	modified:   backend/api/tests/routes/modules/assignments/mark_allocator/put_test.rs 	modified:   backend/api/tests/routes/modules/assignments/statistics/get_test.rs 	modified:   backend/api/tests/routes/modules/assignments/tasks/get_test.rs 	modified:   backend/marker/src/comparators/percentage_comparator.rs 	modified:   backend/marker/src/comparators/regex_comparator.rs 	modified:   backend/marker/src/lib.rs

+++ backend/marker/src/comparators/percentage_comparator.rs
@@ -72,11 +72,11 @@ impl OutputComparator for PercentageComparator {
             missed_patterns.push(memo_lines[i].clone());
         }
 
-        let percentage = (matched_count / memo_lines.len()) as f64;
+        let percentage = matched_count as f64 / memo_lines.len() as f64;
         let mut awarded = section.value * percentage;
 
         if student_lines.len() > memo_lines.len() && !student_lines.is_empty() {
-            let penalty = (memo_lines.len() / student_lines.len()) as f64;
+            let penalty = memo_lines.len() as f64 / student_lines.len() as f64;
             awarded = awarded * penalty;
         }
